+++ include/Log.h
@@ -47,8 +47,8 @@ class Log {
 
 class LogLevel {
   public:
-    LogLevel(int l) : level(l) { printBuffer << std::boolalpha; }
-    LogLevel(LogLevel&& l) : level(l.level) { }
+    explicit LogLevel(int l) : level(l) { printBuffer << std::boolalpha; }
+    LogLevel(LogLevel&& l) : level(l.level) { printBuffer << std::boolalpha; }
 
     LogLevel& operator<< (const glm::vec2& v) {
         return (*this) << v.x << " " << v.y;